pyproject.toml
[tool.poetry]
name = "updatetaxibot"
version = "0.1.0"
description = ""
authors = ["frank <frank.danil704@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.6.0"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
